fix(canary): settings pressure layout
