+++ clients/packages/canary-client/src/components/ButtonStyled.tsx
@@ -1,0 +1,10 @@
+import { Button } from "bonde-components";
+import styled from "styled-components";
+
+const ButtonStyled = styled(Button)`
+  width: auto;
+  min-width: 140px;
+  margin-right: 15px;
+`;
+
+export default ButtonStyled;
