+++ clients/packages/canary-client/src/components/SlateEditor/index.tsx
@@ -36,8 +36,8 @@ const Styles = styled.div`
 
   .editor--content {
     display: grid;
-    grid-column-start: 2;
-    grid-column-end: 2;
+    grid-column-start: 3;
+    grid-column-end: 3;
     grid-row-start: 1;
     grid-row-end: 3;
 
@@ -52,7 +52,7 @@ const Styles = styled.div`
   .editor--toolbar {
     display: grid;
     grid-column-start: 1;
-    grid-column-end: 1;
+    grid-column-end: 2;
 
     &.buttons {
       grid-template-columns: repeat(3,45px) 135px;
