+++ clients/packages/canary-client/src/scenes/WidgetActions/Settings/Autofire/index.tsx
@@ -2,16 +2,16 @@ import React from "react";
 import {
   InputField,
   TextareaField,
-  Card,
   Header,
-  Validators,
-  Button
+  Validators
 } from "bonde-components";
 import { useTranslation } from "react-i18next";
+import { Container, Row, Col } from 'react-grid-system';
+import ButtonStyled from "../../../../components/ButtonStyled";
+import Panel from '../../../../components/Panel';
 import { Widget } from "../../FetchWidgets";
 import { noSpecialCharacters } from "../../../../services/utils";
 import SettingsForm from '../SettingsForm';
-import { Row } from 'react-grid-system';
 
 type Props = {
   widget: Widget;
@@ -29,55 +29,59 @@ const AutofireForm = ({ widget }: Props): React.ReactElement => {
         settings: widget.settings
       }}
     >
-      {() => (
-        <>
-
-
-          <Card padding={{ x: 40, y: 30 }}>
-            <Header.H3>
-              {t("settings.autofire.title")}
-            </Header.H3>
-            <InputField
-              label={t("settings.autofire.label.sendersName")}
-              name="settings.sender_name"
-              placeholder={t("settings.autofire.placeholder.sendersName")}
-              validate={composeValidators(
-                required(t("settings.autofire.validators.required")),
-                noSpecialCharacters(
-                  t("settings.autofire.validators.noSpecialCharacter")
-                )
-              )}
-            />
-            <InputField
-              label={t("settings.autofire.label.sendersEmail")}
-              name="settings.sender_email"
-              placeholder={t("settings.autofire.placeholder.sendersEmail")}
-              validate={composeValidators(
-                required(t("settings.autofire.validators.required")),
-                isEmail(t("settings.autofire.validators.isEmail"))
-              )}
-            />
-            <InputField
-              label={t("settings.autofire.label.emailSubject")}
-              name="settings.email_subject"
-              placeholder={t("settings.autofire.placeholder.emailSubject")}
-              validate={composeValidators(
-                required(t("settings.autofire.validators.required"))
-              )}
-            />
-            <TextareaField
-              label={t("settings.autofire.label.emailBody")}
-              name="settings.email_text"
-              placeholder={t("settings.autofire.placeholder.emailBody")}
-              validate={composeValidators(
-                required(t("settings.autofire.validators.required"))
-              )}
-            />
-            <Row justify='end'>
-              <Button type='submit'> Salvar alterações</Button>
-            </Row>
-          </Card>
-        </>
+      {({ submitting, dirty }: any) => (
+        <Container fluid style={{ width: "100%", padding: "0" }}>
+          <Row>
+            <Col sm={12} md={12} lg={6}>
+              <Panel>
+                <div style={{ marginBottom: "25px" }}>
+                  <Header.H3>
+                    {t("settings.autofire.title")}
+                  </Header.H3>
+                </div>
+                <InputField
+                  label={t("settings.autofire.label.sendersName")}
+                  name="settings.sender_name"
+                  placeholder={t("settings.autofire.placeholder.sendersName")}
+                  validate={composeValidators(
+                    required(t("settings.autofire.validators.required")),
+                    noSpecialCharacters(
+                      t("settings.autofire.validators.noSpecialCharacter")
+                    )
+                  )}
+                />
+                <InputField
+                  label={t("settings.autofire.label.sendersEmail")}
+                  name="settings.sender_email"
+                  placeholder={t("settings.autofire.placeholder.sendersEmail")}
+                  validate={composeValidators(
+                    required(t("settings.autofire.validators.required")),
+                    isEmail(t("settings.autofire.validators.isEmail"))
+                  )}
+                />
+                <InputField
+                  label={t("settings.autofire.label.emailSubject")}
+                  name="settings.email_subject"
+                  placeholder={t("settings.autofire.placeholder.emailSubject")}
+                  validate={composeValidators(
+                    required(t("settings.autofire.validators.required"))
+                  )}
+                />
+                <TextareaField
+                  label={t("settings.autofire.label.emailBody")}
+                  name="settings.email_text"
+                  placeholder={t("settings.autofire.placeholder.emailBody")}
+                  validate={composeValidators(
+                    required(t("settings.autofire.validators.required"))
+                  )}
+                />
+                <Row justify='end'>
+                  <ButtonStyled disabled={submitting || !dirty} type='submit'>{t('settings.defaultForm.submit')}</ButtonStyled>
+                </Row>
+              </Panel>
+            </Col>
+          </Row>
+        </Container>
       )}
     </SettingsForm>
   );
