+++ clients/packages/canary-client/src/scenes/WidgetActions/Settings/ConfigurePostAction/index.tsx
@@ -1,13 +1,16 @@
 import React from "react";
+import { Header, TextareaField } from "bonde-components"
 import { useTranslation } from "react-i18next";
+import { Container, Row, Col } from 'react-grid-system';
+
 import { Widget } from "../../FetchWidgets";
+import ButtonStyled from "../../../../components/ButtonStyled";
 import SpyField from "../../../../components/SpyField";
 import RadioField, { Radio } from "../../../../components/Radio";
 import Panel from "../../../../components/Panel";
 import SettingsForm from '../SettingsForm';
 import DefaultPostAction from "./DefaultPostAction";
 import RichInputField from "./RichInputField";
-import { Header, Button } from "bonde-components"
 
 type Props = {
   widget: Widget;
@@ -34,26 +37,46 @@ const ConfigurePostAction = ({ widget }: Props): React.ReactElement => {
         }
       }}
     >
-      {() => (
+      {({ submitting, dirty }: any) => (
         <Panel>
-          <Header.H3>Pós-ação</Header.H3>
-          <RadioField name='settings.finish_message_type' label={t("settings.finish.title")}>
-            <Radio value='share'>
-              {t("settings.finish.radio.share")}
-            </Radio>
-            <Radio value='custom'>
-              {t("settings.finish.radio.custom")}
-            </Radio>
-          </RadioField>
-
+          <Container fluid style={{ width: "100%", padding: "0" }}>
           <SpyField field='settings.finish_message_type'>
-            {({ value }: any) => value === 'share'
-              ? <DefaultPostAction />
-              : <RichInputField name='settings.finish_message' />
-            }
-          </SpyField>
-
-          <Button type='submit'> Salvar alterações</Button>
+            {({ value }) => (
+              <Row justify="between">
+                <Col sm={12} md={12} lg={6}>
+                  <div style={{ display: "flex", flexDirection: "column" }}>
+                    <div style={{ marginBottom: "25px" }}>
+                      <Header.H3>Pós-ação</Header.H3>
+                    </div>
+                    <RadioField name='settings.finish_message_type' label={t("settings.finish.title")}>
+                      <Radio value='share'>
+                        {t("settings.finish.radio.share")}
+                      </Radio>
+                      <Radio value='custom'>
+                        {t("settings.finish.radio.custom")}
+                      </Radio>
+                    </RadioField>
+                    {value === 'share' ? (
+                      <TextareaField
+                        label={t("settings.finish.default.whatsapp.label")}
+                        name="settings.whatsapp_text"
+                        placeholder={t("settings.finish.default.whatsapp.placeholder")}
+                      />
+                    ) : (
+                      <RichInputField name='settings.finish_message' />
+                    )}
+                    <Row justify='end'>
+                      <ButtonStyled disabled={submitting || !dirty} type='submit'>{t('settings.defaultForm.submit')}</ButtonStyled>
+                    </Row>
+                  </div>
+                </Col>
+                <Col sm={12} md={12} lg={5}>
+                  {value === 'share' && <DefaultPostAction />}
+                </Col>
+              </Row>
+            )}
+            </SpyField>
+          </Container>
         </Panel>
       )}
     </SettingsForm>
